add skewtest
