+++ README.md
@@ -26,6 +26,9 @@ Statistical tests on xarray objects. Currently includes the following statistica
 - `xstatstests.fligner` : Fligner-Killeen test for the variances of K independent samples
 - `xstatstests.median_test` : Mood test for the medians of K independent samples
 
+#### Normality tests
+- `xstatstests.skewtest` : D’Agostino test for normal skewness
+
 #### Other tests
 - `xstatstests.mannwhitneyu` : Mann-Whitney U rank test on two independent samples
 - `xstatstests.ranksums` : Wilcoxon rank-sum statistic on two independent samples
